LPD-84341 Decouple accounts data set into shared component
