+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/contacts/pages/account/List.tsx
@@ -1,57 +1,19 @@
 import * as API from 'shared/api';
 import * as breadcrumbs from 'shared/util/breadcrumbs';
+import AccountsDataSet from 'shared/components/AccountsDataSet';
 import BasePage from 'shared/components/base-page';
-import Card from 'shared/components/Card';
 import Link from '@clayui/link';
 import Loading from 'shared/components/Loading';
 import NoResultsDisplay from 'shared/components/NoResultsDisplay';
 import React from 'react';
 import TotalAccounts from 'contacts/components/account/TotalAccounts';
 import URLConstants from 'shared/util/url-constants';
-import {columns, pagination, useSnapshots} from 'shared/util/frontend-data-set';
 import {isNil} from 'lodash/fp';
-import {LifecycleStages} from './utils/constants';
 import {Routes, toRoute} from 'shared/util/router';
 import {Sizes} from 'shared/util/constants';
-import {toThousands} from 'shared/util/numbers';
 import {useChannelContext} from 'shared/context/channel';
 import {useCurrentUser} from 'shared/hooks/useCurrentUser';
-import {useFrontendDataSet} from 'shared/hooks/useFrontendDataSet';
 import {useRequest} from 'shared/hooks/useRequest';
-
-const lifecycleStagesLabelMap = {
-	[LifecycleStages.AT_RISK]: {
-		displayType: 'danger',
-		label: Liferay.Language.get('at-risk')
-	},
-	[LifecycleStages.AWARE]: {
-		displayType: 'secondary',
-		label: Liferay.Language.get('aware')
-	},
-	[LifecycleStages.ENGAGED]: {
-		displayType: 'warning',
-		label: Liferay.Language.get('engaged')
-	},
-	[LifecycleStages.ESTABLISHED]: {
-		displayType: 'success',
-		label: Liferay.Language.get('established')
-	},
-	[LifecycleStages.ONBOARDING]: {
-		displayType: 'secondary',
-		label: Liferay.Language.get('onboarding')
-	},
-	[LifecycleStages.PIPELINE]: {
-		displayType: 'info',
-		label: Liferay.Language.get('pipeline')
-	}
-};
-
-const lifecycleStageFilter = {
-	items: Object.entries(lifecycleStagesLabelMap).map(([stage]) => ({
-		label: lifecycleStagesLabelMap[stage as LifecycleStages].label,
-		value: stage
-	}))
-};
 
 interface IListProps {
 	channelId: string;
@@ -125,10 +87,6 @@ const List: React.FC<IListProps> = ({channelId, groupId}) => {
 		/>
 	);
 
-	const FrontendDataSet = useFrontendDataSet();
-
-	const snapshots = useSnapshots('accounts-list-dataset');
-
 	if (dataSourceLoading) {
 		return <Loading />;
 	}
@@ -156,192 +114,11 @@ const List: React.FC<IListProps> = ({channelId, groupId}) => {
 					<>
 						<TotalAccounts groupId={groupId} />
 
-						{FrontendDataSet && (
-							<Card>
-								<FrontendDataSet
-									apiURL={`/o/faro/contacts/${groupId}/account/search`}
-									configInURLBehavior='off'
-									customDataRenderers={{
-										accountLifecycleStageRenderer: ({
-											value
-										}) =>
-											value &&
-											columns.cmsLabelRenderer({
-												displayType:
-													lifecycleStagesLabelMap[
-														value
-													].displayType,
-												label:
-													lifecycleStagesLabelMap[
-														value
-													].label
-											}),
-										accountNameRenderer: ({
-											itemData,
-											value
-										}) =>
-											columns.nameAndLinkRenderer({
-												groupId,
-												itemData,
-												route: Routes.CONTACTS_ACCOUNT,
-												value
-											}),
-										annualRevenueRenderer: ({value}) => (
-											<div>{toThousands(value)}</div>
-										),
-										dateRenderer: ({value}) =>
-											columns.dateRenderer({value})
-									}}
-									emptyState={{
-										description: Liferay.Language.get(
-											'no-accounts-were-synced-from-the-connected-data-sources'
-										),
-										image: '/states/satellite.svg',
-										title: Liferay.Language.get(
-											'no-accounts-found'
-										)
-									}}
-									filters={[
-										{
-											id: 'lifecycleStatus',
-											items: lifecycleStageFilter.items,
-											label: Liferay.Language.get(
-												'status'
-											),
-											name: 'status',
-											type: 'selection'
-										},
-										{
-											apiURL: `/o/faro/contacts/${groupId}/account/fds_field_values?channelId=${channelId}&fieldMappingFieldName=industry`,
-											entityFieldType: 'string',
-											id: 'industry',
-											itemKey: 'name',
-											itemLabel: 'name',
-											label: Liferay.Language.get(
-												'industry'
-											),
-											multiple: true,
-											type: 'selection'
-										},
-										{
-											apiURL: `/o/faro/contacts/${groupId}/account/fds_field_values?channelId=${channelId}&fieldMappingFieldName=country`,
-											entityFieldType: 'string',
-											id: 'country',
-											itemKey: 'name',
-											itemLabel: 'name',
-											label: Liferay.Language.get(
-												'country'
-											),
-											multiple: true,
-											type: 'selection'
-										}
-									]}
-									id='accounts-list-dataset'
-									loading={dataSourceLoading}
-									pagination={pagination}
-									showPagination
-									snapshots={snapshots}
-									snapshotsEnabled
-									sort={[
-										{
-											active: true,
-											direction: 'asc',
-											key: 'accountName',
-											label: Liferay.Language.get(
-												'account'
-											)
-										}
-									]}
-									views={[
-										{
-											contentRenderer: 'table',
-											default: true,
-											label: Liferay.Language.get(
-												'default-view'
-											),
-											name: 'table',
-											schema: {
-												fields: [
-													{
-														_key: 'accountName',
-														contentRenderer:
-															'accountNameRenderer',
-														fieldName:
-															'accountName',
-														label: Liferay.Language.get(
-															'account'
-														),
-														sortable: true,
-														truncate: true
-													},
-													{
-														_key: 'industry',
-														fieldName: 'industry',
-														label: Liferay.Language.get(
-															'industry'
-														),
-														sortable: true
-													},
-													{
-														_key: 'lifecycleStage',
-														contentRenderer:
-															'accountLifecycleStageRenderer',
-														fieldName:
-															'lifecycleStage',
-														label: Liferay.Language.get(
-															'lifecycle-stage'
-														),
-														sortable: true
-													},
-													{
-														_key: 'annualRevenue',
-														contentRenderer:
-															'annualRevenueRenderer',
-														fieldName:
-															'annualRevenue',
-														label: Liferay.Language.get(
-															'annual-revenue'
-														),
-														sortable: true
-													},
-													{
-														_key: 'country',
-														fieldName: 'country',
-														label: Liferay.Language.get(
-															'country'
-														),
-														sortable: true
-													},
-													{
-														_key: 'lastActive',
-														contentRenderer:
-															'dateRenderer',
-														fieldName: 'lastActive',
-														label: Liferay.Language.get(
-															'last-active'
-														),
-														sortable: true
-													},
-													{
-														_key: 'lastEnriched',
-														contentRenderer:
-															'dateRenderer',
-														fieldName:
-															'lastEnriched',
-														label: Liferay.Language.get(
-															'last-enriched'
-														),
-														sortable: true,
-														visible: false
-													}
-												]
-											},
-											thumbnail: 'table'
-										}
-									]}
-								/>
-							</Card>
-						)}
+						<AccountsDataSet
+							channelId={channelId}
+							groupId={groupId}
+							loading={dataSourceLoading}
+						/>
 					</>
 				) : (
 					<NoDataSourcesConnected />
