+++ modules/dxp/apps/osb/osb-faro/osb-faro-web/src/main/js/shared/components/AccountsDataSet.tsx
@@ -1,0 +1,211 @@
+import Card from 'shared/components/Card';
+import React from 'react';
+import {columns, pagination, useSnapshots} from 'shared/util/frontend-data-set';
+import {LifecycleStages} from 'contacts/pages/account/utils/constants';
+import {Routes} from 'shared/util/router';
+import {toThousands} from 'shared/util/numbers';
+import {useFrontendDataSet} from 'shared/hooks/useFrontendDataSet';
+
+const lifecycleStagesLabelMap = {
+	[LifecycleStages.AT_RISK]: {
+		displayType: 'danger',
+		label: Liferay.Language.get('at-risk')
+	},
+	[LifecycleStages.AWARE]: {
+		displayType: 'secondary',
+		label: Liferay.Language.get('aware')
+	},
+	[LifecycleStages.ENGAGED]: {
+		displayType: 'warning',
+		label: Liferay.Language.get('engaged')
+	},
+	[LifecycleStages.ESTABLISHED]: {
+		displayType: 'success',
+		label: Liferay.Language.get('established')
+	},
+	[LifecycleStages.ONBOARDING]: {
+		displayType: 'secondary',
+		label: Liferay.Language.get('onboarding')
+	},
+	[LifecycleStages.PIPELINE]: {
+		displayType: 'info',
+		label: Liferay.Language.get('pipeline')
+	}
+};
+
+const lifecycleStageFilter = {
+	items: Object.entries(lifecycleStagesLabelMap).map(([stage]) => ({
+		label: lifecycleStagesLabelMap[stage as LifecycleStages].label,
+		value: stage
+	}))
+};
+
+interface IAccountsDataSetProps {
+	channelId: string;
+	groupId: string;
+	loading?: boolean;
+}
+
+const AccountsDataSet: React.FC<IAccountsDataSetProps> = ({
+	channelId,
+	groupId,
+	loading
+}) => {
+	const FrontendDataSet = useFrontendDataSet();
+
+	const snapshots = useSnapshots('accounts-list-dataset');
+
+	if (!FrontendDataSet) {
+		return null;
+	}
+
+	return (
+		<Card>
+			<FrontendDataSet
+				apiURL={`/o/faro/contacts/${groupId}/account/search`}
+				configInURLBehavior='off'
+				customDataRenderers={{
+					accountLifecycleStageRenderer: ({value}) =>
+						value &&
+						columns.cmsLabelRenderer({
+							displayType:
+								lifecycleStagesLabelMap[value].displayType,
+							label: lifecycleStagesLabelMap[value].label
+						}),
+					accountNameRenderer: ({itemData, value}) =>
+						columns.nameAndLinkRenderer({
+							groupId,
+							itemData,
+							route: Routes.CONTACTS_ACCOUNT,
+							value
+						}),
+					annualRevenueRenderer: ({value}) => (
+						<div>{toThousands(value)}</div>
+					),
+					dateRenderer: ({value}) => columns.dateRenderer({value})
+				}}
+				emptyState={{
+					description: Liferay.Language.get(
+						'no-accounts-were-synced-from-the-connected-data-sources'
+					),
+					image: '/states/satellite.svg',
+					title: Liferay.Language.get('no-accounts-found')
+				}}
+				filters={[
+					{
+						id: 'lifecycleStatus',
+						items: lifecycleStageFilter.items,
+						label: Liferay.Language.get('status'),
+						name: 'status',
+						type: 'selection'
+					},
+					{
+						apiURL: `/o/faro/contacts/${groupId}/account/fds_field_values?channelId=${channelId}&fieldMappingFieldName=industry`,
+						entityFieldType: 'string',
+						id: 'industry',
+						itemKey: 'name',
+						itemLabel: 'name',
+						label: Liferay.Language.get('industry'),
+						multiple: true,
+						type: 'selection'
+					},
+					{
+						apiURL: `/o/faro/contacts/${groupId}/account/fds_field_values?channelId=${channelId}&fieldMappingFieldName=country`,
+						entityFieldType: 'string',
+						id: 'country',
+						itemKey: 'name',
+						itemLabel: 'name',
+						label: Liferay.Language.get('country'),
+						multiple: true,
+						type: 'selection'
+					}
+				]}
+				id='accounts-list-dataset'
+				loading={loading}
+				pagination={pagination}
+				showPagination
+				snapshots={snapshots}
+				snapshotsEnabled
+				sort={[
+					{
+						active: true,
+						direction: 'asc',
+						key: 'accountName',
+						label: Liferay.Language.get('account')
+					}
+				]}
+				views={[
+					{
+						contentRenderer: 'table',
+						default: true,
+						label: Liferay.Language.get('default-view'),
+						name: 'table',
+						schema: {
+							fields: [
+								{
+									_key: 'accountName',
+									contentRenderer: 'accountNameRenderer',
+									fieldName: 'accountName',
+									label: Liferay.Language.get('account'),
+									sortable: true,
+									truncate: true
+								},
+								{
+									_key: 'industry',
+									fieldName: 'industry',
+									label: Liferay.Language.get('industry'),
+									sortable: true
+								},
+								{
+									_key: 'lifecycleStage',
+									contentRenderer:
+										'accountLifecycleStageRenderer',
+									fieldName: 'lifecycleStage',
+									label: Liferay.Language.get(
+										'lifecycle-stage'
+									),
+									sortable: true
+								},
+								{
+									_key: 'annualRevenue',
+									contentRenderer: 'annualRevenueRenderer',
+									fieldName: 'annualRevenue',
+									label: Liferay.Language.get(
+										'annual-revenue'
+									),
+									sortable: true
+								},
+								{
+									_key: 'country',
+									fieldName: 'country',
+									label: Liferay.Language.get('country'),
+									sortable: true
+								},
+								{
+									_key: 'lastActive',
+									contentRenderer: 'dateRenderer',
+									fieldName: 'lastActive',
+									label: Liferay.Language.get('last-active'),
+									sortable: true
+								},
+								{
+									_key: 'lastEnriched',
+									contentRenderer: 'dateRenderer',
+									fieldName: 'lastEnriched',
+									label: Liferay.Language.get(
+										'last-enriched'
+									),
+									sortable: true,
+									visible: false
+								}
+							]
+						},
+						thumbnail: 'table'
+					}
+				]}
+			/>
+		</Card>
+	);
+};
+
+export default AccountsDataSet;
